test(M-54): канарейка T4b мерила не тот периметр + clippy в собственных оракулах
Две мои ошибки, вскрытые первым прогоном гейта (VERDICT: FAIL, 2 нарушения).

1. T4b грепала ВЕСЬ gateway-serve/src/lib.rs на snapshot_from_checkpoint и краснела на
   корректной реализации. Символ обязан остаться в крейте: на нём стоит passthrough-адаптер
   serve::snapshot_msg, который оракулы M-46 (red_ws_series_vs_replay) используют как
   НЕЗАВИСИМЫЙ эталон сверки WS<->реплей. Проверять надо путь подключения, а не присутствие
   символа в файле. Теперь канарейка извлекает тело run_authorized_session (без комментариев)
   и проверяет три вещи: второго прохода на пути подключения нет; снапшот берётся из
   live.snapshot(); эталон snapshot_from_checkpoint в крейте СОХРАНЁН (обратная канарейка —
   иначе кто-нибудь 'почистит' его заодно с оптимизацией и сверка станет тавтологией).

2. clippy -D warnings краснел на 'events_decoded as u64' в моих же оракулах — поле уже u64.
   Паритет verify<->CI, который я сам записал в gates.md §3, нарушен мной же.

Проверено фактически: run_authorized_session идёт resume -> pump -> ServeMsg::Snapshot(
live.snapshot()); snapshot_from_checkpoint остался только в serve::snapshot_msg и в
комментарии. Реализация dev'а верна, красной была канарейка.

verify_M-54.sh: 12/12 PASS, exit=0.

+++ crates/gateway/tests/red_connect_cost_single.rs
@@ -149,7 +149,7 @@ fn pump_to_tail(live: &mut gateway::LiveReducer, dir: &std::path::Path) -> u64 {
         let (frames, _cursor, stats) = live
             .pump(dir, EpochFilter::OwnCaptureOnly, 128)
             .expect("pump");
-        decoded += stats.events_decoded as u64;
+        decoded += stats.events_decoded;
         if frames.is_empty() {
             break;
         }
